requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
-e git+git@github.com:bsmiller25/espn-api.git@7ce7617816b660ddee4aaf4cacb22c1fbb01d47f#egg=espn_api
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
numpy==1.17.3
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.12.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
